+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 master
 ------
 
+- (`#110 <https://github.com/GranthamImperial/silicone/pull/108>`_) Gave an option to time_dep_ratio to prevent using any variables that appear only at certain times.
 - (`#108 <https://github.com/GranthamImperial/silicone/pull/108>`_) Added a multiple infiller to split up an aggregate with a remainder. Disabled test for downloading database.
 - (`#103 <https://github.com/GranthamImperial/silicone/pull/103>`_) Update github address to GranthamImperial.
 - (`#101 <https://github.com/GranthamImperial/silicone/pull/101>`_) Update release docs